chore: extract charts shared/loaders/equity/quality from dashboard index (move-only)
