+++ src/routes/dashboard/charts/equity.ts
@@ -1,0 +1,98 @@
+import { or } from 'drizzle-orm'
+import { type ChartsBodyOverview, ECHARTS_CDN } from './shared'
+import { safeJsonScript } from '../shared'
+
+export interface EquityPoint {
+  date: string // YYYY-MM-DD (JST)
+  dailyPnl: number
+  cumulativePnl: number
+  drawdownPct: number // 0 or 負 (peak からの低下率)
+}
+
+/**
+ * trade_journal の post_submit で realized_pnl が記録されている SELL fill を
+ * 日次集計し、累積 PnL とドローダウン率を計算する。
+ *
+ * - peak は cumulativePnl の rolling max
+ * - drawdownPct は peak が 0 以下のとき null 相当 (= 0%) として扱う
+ *   (peak が小さい初期は割り算が暴れるため)
+ */
+export async function loadEquityCurve(db: D1Database): Promise<EquityPoint[]> {
+  // SQLite の date(timestamp) はデフォルト UTC。JST 表示にするため +9h shift。
+  const result = await db
+    .prepare(
+      `SELECT date(timestamp, '+9 hours') AS day,
+              SUM(realized_pnl) AS daily_pnl
+       FROM trade_journal
+       WHERE realized_pnl IS NOT NULL
+         AND trade_event_type = 'post_submit'
+       GROUP BY day
+       ORDER BY day ASC`,
+    )
+    .all<{ day: string; daily_pnl: number }>()
+  const rows = result.results ?? []
+  return computeEquitySeries(rows.map((r) => ({ date: r.day, dailyPnl: Number(r.daily_pnl) })))
+}
+
+export function computeEquitySeries(
+  daily: Array<{ date: string; dailyPnl: number }>,
+): EquityPoint[] {
+  const points: EquityPoint[] = []
+  let cumulative = 0
+  let peak = 0
+  for (const d of daily) {
+    cumulative += d.dailyPnl
+    if (cumulative > peak) peak = cumulative
+    // peak が +側になるまでは drawdown を 0 表示 (シード資金規模が不明なので
+    // % 計算は意味をなさない。peak を絶対額として比較するのも手だが、トレーダー
+    // 視点では「最高益からの下落率」が読みたいので peak>0 で初めて非ゼロに)
+    const dd = peak > 0 ? (cumulative - peak) / peak : 0
+    points.push({ date: d.date, dailyPnl: d.dailyPnl, cumulativePnl: cumulative, drawdownPct: dd })
+  }
+  return points
+}
+
+export function renderOverviewTab(args: ChartsBodyOverview): string {
+  if (args.equity.length === 0) {
+    return `<p class="muted">まだ実 fill (realized_pnl) が無いためエクイティカーブを描けません。最初の SELL が約定すると表示されます。</p>`
+  }
+  const initScript = `
+    document.addEventListener('DOMContentLoaded', function () {
+      if (typeof echarts === 'undefined') return;
+      var data = window.__chartData;
+      var dates = data.equity.map(function (p) { return p.date; });
+      var equity = data.equity.map(function (p) { return p.cumulativePnl; });
+      var dd = data.equity.map(function (p) { return p.drawdownPct * 100; });
+      var equityChart = echarts.init(document.getElementById('equity-chart'));
+      equityChart.setOption({
+        title: { text: '累積 realized PnL', left: 'center', textStyle: { fontSize: 14 } },
+        tooltip: { trigger: 'axis', valueFormatter: function (v) { return Number(v).toFixed(2); } },
+        grid: { left: 50, right: 20, top: 40, bottom: 40 },
+        xAxis: { type: 'category', data: dates },
+        yAxis: { type: 'value', name: 'PnL', axisLabel: { formatter: '{value}' } },
+        series: [{ type: 'line', data: equity, smooth: false, areaStyle: { opacity: 0.1 }, lineStyle: { width: 2 } }],
+      });
+      var ddChart = echarts.init(document.getElementById('dd-chart'));
+      ddChart.setOption({
+        title: { text: 'ドローダウン (累積 PnL の peak からの低下率)', left: 'center', textStyle: { fontSize: 14 } },
+        tooltip: { trigger: 'axis', valueFormatter: function (v) { return Number(v).toFixed(2) + '%'; } },
+        grid: { left: 50, right: 20, top: 40, bottom: 40 },
+        xAxis: { type: 'category', data: dates },
+        yAxis: { type: 'value', max: 0, axisLabel: { formatter: '{value}%' } },
+        series: [{ type: 'line', data: dd, areaStyle: { color: '#c22', opacity: 0.2 }, lineStyle: { color: '#c22', width: 1 } }],
+      });
+      window.addEventListener('resize', function () { equityChart.resize(); ddChart.resize(); });
+    });
+  `
+  return `<p class="muted" style="font-size:12px">
+    累積 realized PnL と peak からの下落率 (MaxDD)。戦略の長期パフォーマンス指標。
+    シード資金額を保持していないため下落率は「累積 PnL の peak からの相対」で計算
+    (peak ≤ 0 のときは 0%)。当日 intraday の risk halt 閾値 (drawdown_kill /
+    risk_dd_halt) は別概念のため重畳しない。
+  </p>
+  <div id="equity-chart" style="width:100%;height:320px;background:#fff;border:1px solid #d0d0d5;border-radius:6px;margin-top:12px"></div>
+  <div id="dd-chart" style="width:100%;height:280px;background:#fff;border:1px solid #d0d0d5;border-radius:6px;margin-top:12px"></div>
+  ${safeJsonScript('__chartData', { equity: args.equity })}
+  <script src="${ECHARTS_CDN}" defer></script>
+  <script>${initScript}</script>`
+}
