+++ src/routes/dashboard/charts/quality.ts
@@ -1,0 +1,245 @@
+import { type ChartsBodyQuality, ECHARTS_CDN } from './shared'
+import { safeJsonScript } from '../shared'
+
+/**
+ * Decision breakdown chart 用の日次集計 (#158 Phase 2)。
+ *
+ * strategy_decision_log を JST 日次でグルーピングし、各 decision
+ * (BUY/SELL/HOLD/SKIP/REJECT/ERROR) のカウントを返す。トレーダーは
+ * 「BUY/SELL が出すぎ・出なさすぎ」「SKIP/REJECT が偏ってないか」を一目で
+ * 見たいので、1 日 1 行 × 6 系列の stacked bar 用のデータ形にする。
+ *
+ * 直近 90 日のみ (それ以上はチャートが詰まって読めない)。
+ */
+export const DECISION_KEYS = ['BUY', 'SELL', 'HOLD', 'SKIP', 'REJECT', 'ERROR'] as const
+
+export type DecisionKey = (typeof DECISION_KEYS)[number]
+
+export interface DecisionBreakdownPoint {
+  date: string
+  counts: Record<DecisionKey, number>
+}
+
+export async function loadDecisionBreakdown(db: D1Database): Promise<DecisionBreakdownPoint[]> {
+  const result = await db
+    .prepare(
+      `SELECT date(timestamp, '+9 hours') AS day,
+              decision,
+              COUNT(*) AS n
+       FROM strategy_decision_log
+       WHERE timestamp >= date('now', '-90 days')
+       GROUP BY day, decision
+       ORDER BY day ASC`,
+    )
+    .all<{ day: string; decision: string; n: number }>()
+  return aggregateDecisionRows(result.results ?? [])
+}
+
+export function aggregateDecisionRows(
+  rows: Array<{ day: string; decision: string; n: number }>,
+): DecisionBreakdownPoint[] {
+  const map = new Map<string, Record<DecisionKey, number>>()
+  for (const r of rows) {
+    let bucket = map.get(r.day)
+    if (!bucket) {
+      bucket = { BUY: 0, SELL: 0, HOLD: 0, SKIP: 0, REJECT: 0, ERROR: 0 }
+      map.set(r.day, bucket)
+    }
+    // 想定外 decision (将来追加など) は ERROR バケットに寄せて見落とし防止
+    const key: DecisionKey = (DECISION_KEYS as readonly string[]).includes(r.decision)
+      ? (r.decision as DecisionKey)
+      : 'ERROR'
+    bucket[key] += Number(r.n)
+  }
+  return [...map.entries()]
+    .sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0))
+    .map(([date, counts]) => ({ date, counts }))
+}
+
+/**
+ * Per-trade realized PnL を全件取得 (#158 Phase 3)。
+ *
+ * trade_journal.realized_pnl は SELL fill に確定損益が記録される (BUY は null)。
+ * 戦略のエッジが本物か偽物かを見るために、勝率・profit factor・expectancy を
+ * 計算 + 分布を histogram で可視化する。
+ */
+export async function loadTradePnls(db: D1Database): Promise<number[]> {
+  const result = await db
+    .prepare(
+      `SELECT realized_pnl AS pnl
+       FROM trade_journal
+       WHERE realized_pnl IS NOT NULL
+         AND trade_event_type = 'post_submit'
+       ORDER BY id ASC`,
+    )
+    .all<{ pnl: number }>()
+  return (result.results ?? []).map((r) => Number(r.pnl)).filter((n) => Number.isFinite(n))
+}
+
+export interface TradeStats {
+  count: number
+  wins: number
+  losses: number
+  /** 0..1 (勝率) */
+  winRate: number
+  avgWin: number
+  avgLoss: number // 負値
+  /** 総利益 / |総損失|。loss=0 のときは Infinity (UI 側で "—" 表示) */
+  profitFactor: number
+  /** 1 trade あたり期待損益 = winRate * avgWin + (1-winRate) * avgLoss */
+  expectancy: number
+  total: number
+}
+
+/**
+ * 「エッジが本物か」を 1 表で見るためのサマリ統計。break-even (pnl=0) は wins / losses
+ * どちらにも入れない (エクスペクタンシ計算で 0 として中立に効く)。
+ */
+export function computeTradeStats(pnls: number[]): TradeStats {
+  if (pnls.length === 0) {
+    return { count: 0, wins: 0, losses: 0, winRate: 0, avgWin: 0, avgLoss: 0, profitFactor: 0, expectancy: 0, total: 0 }
+  }
+  let wins = 0
+  let losses = 0
+  let sumWin = 0
+  let sumLoss = 0
+  let total = 0
+  for (const p of pnls) {
+    total += p
+    if (p > 0) {
+      wins += 1
+      sumWin += p
+    } else if (p < 0) {
+      losses += 1
+      sumLoss += p
+    }
+  }
+  const decisive = wins + losses
+  const winRate = decisive > 0 ? wins / decisive : 0
+  const avgWin = wins > 0 ? sumWin / wins : 0
+  const avgLoss = losses > 0 ? sumLoss / losses : 0
+  const profitFactor = sumLoss < 0 ? sumWin / Math.abs(sumLoss) : sumWin > 0 ? Infinity : 0
+  const expectancy = winRate * avgWin + (1 - winRate) * avgLoss
+  return { count: pnls.length, wins, losses, winRate, avgWin, avgLoss, profitFactor, expectancy, total }
+}
+
+export interface PnlHistogramBin {
+  label: string // "(-5, -3]" など
+  binStart: number
+  binEnd: number
+  binCenter: number
+  count: number
+}
+
+/**
+ * pnl 値を最大 12 ビンの histogram に。範囲は対称 (max(|min|, max)) で
+ * 0 を境に正/負で色分け可能にする。サンプルが少なすぎるとビン数を減らす。
+ */
+export function computePnlHistogram(pnls: number[], maxBins = 12): PnlHistogramBin[] {
+  if (pnls.length === 0) return []
+  const absMax = Math.max(...pnls.map(Math.abs))
+  if (absMax === 0) {
+    return [{ label: '0', binStart: 0, binEnd: 0, binCenter: 0, count: pnls.length }]
+  }
+  const bins = Math.min(maxBins, Math.max(3, Math.ceil(Math.sqrt(pnls.length))))
+  const range = absMax * 2
+  const width = range / bins
+  const out: PnlHistogramBin[] = []
+  for (let i = 0; i < bins; i += 1) {
+    const start = -absMax + width * i
+    const end = start + width
+    out.push({
+      label: `(${start.toFixed(1)}, ${end.toFixed(1)}]`,
+      binStart: start,
+      binEnd: end,
+      binCenter: (start + end) / 2,
+      count: 0,
+    })
+  }
+  for (const p of pnls) {
+    // 末尾の bin は閉区間 [end, end] を含むよう特別処理
+    let idx = Math.floor((p - -absMax) / width)
+    if (idx >= bins) idx = bins - 1
+    if (idx < 0) idx = 0
+    out[idx]!.count += 1
+  }
+  return out
+}
+
+export function renderTradeStatsTable(s: TradeStats): string {
+  if (s.count === 0) return ''
+  const fmt = (n: number) => (Number.isFinite(n) ? n.toFixed(2) : '—')
+  const pct = (n: number) => (Number.isFinite(n) ? (n * 100).toFixed(1) + '%' : '—')
+  const pf = Number.isFinite(s.profitFactor) ? s.profitFactor.toFixed(2) : '∞'
+  const expClass = s.expectancy > 0 ? 'ok' : s.expectancy < 0 ? 'err' : 'muted'
+  return `<table style="margin-top:12px">
+    <thead><tr><th>件数</th><th>勝</th><th>負</th><th>勝率</th><th>平均利益</th><th>平均損失</th><th>profit factor</th><th>expectancy / trade</th><th>合計</th></tr></thead>
+    <tbody><tr>
+      <td>${s.count}</td>
+      <td class="ok">${s.wins}</td>
+      <td class="err">${s.losses}</td>
+      <td>${pct(s.winRate)}</td>
+      <td class="ok">${fmt(s.avgWin)}</td>
+      <td class="err">${fmt(s.avgLoss)}</td>
+      <td>${pf}</td>
+      <td class="${expClass}">${fmt(s.expectancy)}</td>
+      <td class="${s.total > 0 ? 'ok' : s.total < 0 ? 'err' : 'muted'}">${fmt(s.total)}</td>
+    </tr></tbody>
+  </table>`
+}
+
+export function renderQualityTab(args: ChartsBodyQuality): string {
+  if (args.pnls.length === 0 && args.decisions.length === 0) {
+    return `<p class="muted">まだ判定ログも実 fill も無いため取引品質を描けません。cron が動き出すと judgement breakdown、SELL が約定すると PnL 分布が出ます。</p>`
+  }
+  const initScript = `
+    document.addEventListener('DOMContentLoaded', function () {
+      if (typeof echarts === 'undefined') return;
+      var data = window.__chartData;
+      var DECISION_KEYS = ['BUY', 'SELL', 'HOLD', 'SKIP', 'REJECT', 'ERROR'];
+      var DECISION_COLORS = { BUY: '#057a55', SELL: '#1471a8', HOLD: '#aaa', SKIP: '#b25000', REJECT: '#7c3aed', ERROR: '#c22' };
+      var dbDates = data.decisions.map(function (p) { return p.date; });
+      var dbEl = document.getElementById('decision-chart');
+      if (dbEl && dbDates.length > 0) {
+        var dbChart = echarts.init(dbEl);
+        dbChart.setOption({
+          title: { text: '日次 Decision breakdown (BUY / SELL / HOLD / SKIP / REJECT / ERROR)', left: 'center', textStyle: { fontSize: 14 } },
+          tooltip: { trigger: 'axis', axisPointer: { type: 'shadow' } },
+          legend: { top: 22 },
+          grid: { left: 50, right: 20, top: 60, bottom: 40 },
+          xAxis: { type: 'category', data: dbDates },
+          yAxis: { type: 'value', name: '判定数' },
+          series: DECISION_KEYS.map(function (k) {
+            return { name: k, type: 'bar', stack: 'decisions',
+              data: data.decisions.map(function (p) { return p.counts[k] || 0; }),
+              itemStyle: { color: DECISION_COLORS[k] } };
+          }),
+        });
+        window.addEventListener('resize', function () { dbChart.resize(); });
+      }
+      var pnlHistEl = document.getElementById('pnl-hist-chart');
+      if (pnlHistEl && data.histogram && data.histogram.length > 0) {
+        var pnlHist = echarts.init(pnlHistEl);
+        pnlHist.setOption({
+          title: { text: 'Per-trade realized PnL 分布', left: 'center', textStyle: { fontSize: 14 } },
+          tooltip: { trigger: 'axis', axisPointer: { type: 'shadow' },
+            formatter: function (params) { var p = params[0]; return p.name + ': ' + p.value + ' trades'; } },
+          grid: { left: 50, right: 20, top: 40, bottom: 40 },
+          xAxis: { type: 'category', data: data.histogram.map(function (b) { return b.label; }) },
+          yAxis: { type: 'value', name: 'trades' },
+          series: [{ type: 'bar',
+            data: data.histogram.map(function (b) {
+              return { value: b.count, itemStyle: { color: b.binCenter >= 0 ? '#057a55' : '#c22' } };
+            }) }],
+        });
+        window.addEventListener('resize', function () { pnlHist.resize(); });
+      }
+    });
+  `
+  return `<div id="pnl-hist-chart" style="width:100%;height:320px;background:#fff;border:1px solid #d0d0d5;border-radius:6px;margin-top:12px"></div>
+  ${renderTradeStatsTable(args.stats)}
+  <div id="decision-chart" style="width:100%;height:320px;background:#fff;border:1px solid #d0d0d5;border-radius:6px;margin-top:12px"></div>
+  ${safeJsonScript('__chartData', { decisions: args.decisions, histogram: args.histogram })}
+  <script src="${ECHARTS_CDN}" defer></script>
+  <script>${initScript}</script>`
+}
